requirements.txt
summa==1.2.0
fuzzywuzzy==0.17.0
lxml==4.4.1
python-docx==0.8.10
python-Levenshtein==0.12.0
selenium==3.141.0
urllib3==1.25.7
